Refactor `target_concurrency` tests
